refactored the numbersRouter and created a numbersController containing the logic used to handle the routed numbers requests.
