+++ src/routes/numbersRouter.js
@@ -1,58 +1,16 @@
 const express = require('express');
-const { add, subtract, multiply, divide, remainder } = require('../lib/numbers');
+const numbersController = require('../controllers/numbersController');
 
 const router = express.Router();
 
-router.get('/add/:firstNum/and/:secondNum', (req, res) => {
-  const firstNum = parseInt(req.params.firstNum);
-  const secondNum = parseInt(req.params.secondNum);
+router.get('/add/:firstNum/and/:secondNum', numbersController.numbers_add);
 
-  return Number.isNaN(firstNum) || Number.isNaN(secondNum)
-    ? res.status(400).json({ error: 'Parameters must be valid numbers.' })
-    : res.status(200).json({ result: add(firstNum, secondNum) });
-});
+router.get('/subtract/:firstNum/from/:secondNum', numbersController.numbers_subtract);
 
-router.get('/subtract/:firstNum/from/:secondNum', (req, res) => {
-  const firstNum = parseInt(req.params.firstNum);
-  const secondNum = parseInt(req.params.secondNum);
+router.post('/multiply', numbersController.numbers_multiply);
 
-  return Number.isNaN(firstNum) || Number.isNaN(secondNum)
-    ? res.status(400).json({ error: 'Parameters must be valid numbers.' })
-    : res.status(200).json({ result: subtract(secondNum, firstNum) });
-});
+router.post('/divide', numbersController.numbers_divide);
 
-router.post('/multiply', (req, res) => {
-  if (req.body.a === undefined || req.body.b === undefined) {
-    res.status(400).json({ error: 'Parameters "a" and "b" are required.' });
-  } else if (Number.isNaN(parseInt(req.body.a)) || Number.isNaN(parseInt(req.body.b))) {
-    res.status(400).json({ error: 'Parameters "a" and "b" must be valid numbers.' });
-  } else {
-    res.status(200).json({ result: multiply(req.body.a, req.body.b) });
-  }
-});
-
-router.post('/divide', (req, res) => {
-  if (req.body.b === 0) {
-    res.status(400).json({ error: 'Unable to divide by 0.' });
-  } else if (req.body.a === undefined || req.body.b === undefined) {
-    res.status(400).json({ error: 'Parameters "a" and "b" are required.' });
-  } else if (Number.isNaN(parseInt(req.body.a)) || Number.isNaN(parseInt(req.body.b))) {
-    res.status(400).json({ error: 'Parameters "a" and "b" must be valid numbers.' });
-  } else {
-    res.status(200).json({ result: divide(req.body.a, req.body.b) });
-  }
-});
-
-router.post('/remainder', (req, res) => {
-  if (req.body.b === 0) {
-    res.status(400).json({ error: 'Unable to divide by 0.' });
-  } else if (req.body.a === undefined || req.body.b === undefined) {
-    res.status(400).json({ error: 'Parameters "a" and "b" are required.' });
-  } else if (Number.isNaN(parseInt(req.body.a)) || Number.isNaN(parseInt(req.body.b))) {
-    res.status(400).json({ error: 'Parameters must be valid numbers.' });
-  } else {
-    res.status(200).json({ result: remainder(req.body.a, req.body.b) });
-  }
-});
+router.post('/remainder', numbersController.numbers_remainder);
 
 module.exports = router;
